Major changes to Macro handling.
1. Macro definitions (name, body, and hash) are now output
as part of the auxiliary database.  All macros in the main
file are output, regardless of if they are actually used.
2. When an AST utilizes a macro defined within the same file, the
hash of the macro is added to the 'macros' field on the AST.  The
hash can then be utilized to find the appropriate element in the
auxiliary database.
3. As before (and as with types), if a macro utilized by an AST
is defined in a seperate header file, the header file is added
to the AST's includes field.