Minor improvements to "includes" field.
This resolves a few lingering issues with the includes
field and system headers.  We now use spelling locations
when performing the include path search, which reduces
the odds of an 'invalid location' cutting the search off
short.  Additionally, using the spelling location helps
ensure the proper header begins the search.